prepare encoder
